+++ data/src/main/java/ua/turskyi/data/database/room/Database.kt
@@ -1,4 +1,4 @@
-package ua.turskyi.data.datastore.room
+package ua.turskyi.data.database.room
 
 import android.content.Context
 import androidx.room.Room.databaseBuilder
@@ -8,7 +8,7 @@ import androidx.sqlite.db.SupportSQLiteDatabase
 import ua.turskyi.data.BuildConfig
 import ua.turskyi.data.entities.local.CityEntity
 import ua.turskyi.data.entities.local.CountryEntity
-import ua.turskyi.data.datastore.room.dao.CountriesDao
+import ua.turskyi.data.database.room.dao.CountriesDao
 import androidx.room.Database as DB
 
 @DB(
