+++ data/src/main/java/ua/turskyi/data/repository/CountriesRepositoryImpl.kt
@@ -6,9 +6,9 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
-import ua.turskyi.data.datastore.room.datasource.CountriesDbSource
+import ua.turskyi.data.database.room.datasource.DatabaseSource
 import ua.turskyi.data.extensions.*
-import ua.turskyi.data.network.datasource.CountriesNetSource
+import ua.turskyi.data.network.datasource.NetSource
 import ua.turskyi.domain.model.CityModel
 import ua.turskyi.domain.model.CountryModel
 import ua.turskyi.domain.repository.CountriesRepository
@@ -16,8 +16,8 @@ import ua.turskyi.domain.repository.CountriesRepository
 class CountriesRepositoryImpl(private val applicationScope: CoroutineScope) : CountriesRepository,
     KoinComponent {
 
-    private val netSource: CountriesNetSource by inject()
-    private val dbSource: CountriesDbSource by inject()
+    private val netSource: NetSource by inject()
+    private val databaseSource: DatabaseSource by inject()
 
     override suspend fun refreshCountries(
         onSuccess: () -> Unit,
@@ -35,8 +35,8 @@ class CountriesRepositoryImpl(private val applicationScope: CoroutineScope) : Co
         onError: ((Exception) -> Unit?)?
     ) {
         applicationScope.launch {
-            dbSource.updateSelfie(id, selfie)
-            onSuccess(dbSource.getVisitedLocalCountriesFromDb().mapEntityListToModelList())
+            databaseSource.updateSelfie(id, selfie)
+            onSuccess(databaseSource.getVisitedLocalCountriesFromDb().mapEntityListToModelList())
         }
     }
 
@@ -49,7 +49,7 @@ class CountriesRepositoryImpl(private val applicationScope: CoroutineScope) : Co
             try {
                 val countryLocal = country.mapModelToEntity()
                 countryLocal.isVisited = true
-                dbSource.insertCountry(countryLocal)
+                databaseSource.insertCountry(countryLocal)
                 onSuccess()
             } catch (exception: Exception) {
                 onError?.invoke(exception)
@@ -66,8 +66,8 @@ class CountriesRepositoryImpl(private val applicationScope: CoroutineScope) : Co
             try {
                 val countryLocal = country.mapModelToEntity()
                 countryLocal.isVisited = false
-                dbSource.removeCitiesByCountry(country.id)
-                dbSource.insertCountry(countryLocal)
+                databaseSource.removeCitiesByCountry(country.id)
+                databaseSource.insertCountry(countryLocal)
                 onSuccess()
             } catch (exception: java.lang.Exception) {
                 onError?.invoke(exception)
@@ -82,7 +82,7 @@ class CountriesRepositoryImpl(private val applicationScope: CoroutineScope) : Co
     ) {
         applicationScope.launch {
             try {
-                dbSource.insertCity(city.mapModelToEntity())
+                databaseSource.insertCity(city.mapModelToEntity())
                 onSuccess.invoke()
             } catch (exception: java.lang.Exception) {
                 onError?.invoke(exception)
@@ -99,7 +99,7 @@ class CountriesRepositoryImpl(private val applicationScope: CoroutineScope) : Co
             try {
                 withContext(Dispatchers.Default) {
                     val cityLocal = city.mapModelToEntity()
-                    dbSource.removeCity(cityLocal)
+                    databaseSource.removeCity(cityLocal)
                 }
                 onSuccess()
             } catch (exception: java.lang.Exception) {
@@ -115,7 +115,7 @@ class CountriesRepositoryImpl(private val applicationScope: CoroutineScope) : Co
     ) {
         applicationScope.launch {
             try {
-                dbSource.insertAllCountries(countries.mapModelListToEntityList())
+                databaseSource.insertAllCountries(countries.mapModelListToEntityList())
                 onSuccess()
             } catch (exception: Exception) {
                 onError?.invoke(exception)
@@ -128,7 +128,7 @@ class CountriesRepositoryImpl(private val applicationScope: CoroutineScope) : Co
         onError: ((Exception) -> Unit?)?
     ) {
         applicationScope.launch {
-            onSuccess(dbSource.getVisitedLocalCountriesFromDb().mapEntityListToModelList())
+            onSuccess(databaseSource.getVisitedLocalCountriesFromDb().mapEntityListToModelList())
         }
     }
 
@@ -136,14 +136,14 @@ class CountriesRepositoryImpl(private val applicationScope: CoroutineScope) : Co
         onSuccess: (List<CityModel>) -> Unit,
         onError: ((Exception) -> Unit?)?
     ) {
-        applicationScope.launch { onSuccess(dbSource.getCities().mapEntitiesToModelList()) }
+        applicationScope.launch { onSuccess(databaseSource.getCities().mapEntitiesToModelList()) }
     }
 
     override suspend fun setCountNotVisitedCountries(
         onSuccess: (Int) -> Unit,
         onError: ((Exception) -> Unit?)?
     ) {
-        applicationScope.launch { onSuccess(dbSource.getCountNotVisitedCountries()) }
+        applicationScope.launch { onSuccess(databaseSource.getCountNotVisitedCountries()) }
     }
 
     override suspend fun getCountriesByRange(
@@ -153,7 +153,7 @@ class CountriesRepositoryImpl(private val applicationScope: CoroutineScope) : Co
         onError: ((Exception) -> Unit?)?
     ) {
         applicationScope.launch {
-            onSuccess(dbSource.getLocalCountriesByRange(to, from).mapEntityListToModelList())
+            onSuccess(databaseSource.getLocalCountriesByRange(to, from).mapEntityListToModelList())
         }
     }
 
@@ -166,7 +166,7 @@ class CountriesRepositoryImpl(private val applicationScope: CoroutineScope) : Co
     ) {
         applicationScope.launch {
             onSuccess(
-                dbSource.loadCountriesByNameAndRange(name, limit, offset).mapEntityListToModelList()
+                databaseSource.loadCountriesByNameAndRange(name, limit, offset).mapEntityListToModelList()
             )
         }
     }
